1570 Move workflow validate and instruction MCP tools to mcp-tool-platform
