--- server/libs/ai/mcp/mcp-tool/mcp-tool-automation/src/main/java/com/bytechef/ai/mcp/tool/automation/model/WorkflowValidationResult.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2025 ByteChef
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.bytechef.ai.mcp.tool.automation.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyDescription;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
-/**
- * Workflow validation result record for API responses.
- *
- * @author Marko Kriskovic
- */
-@SuppressFBWarnings("EI")
-public record WorkflowValidationResult(
-    @JsonProperty("valid") @JsonPropertyDescription("Whether the workflow is valid") boolean valid,
-    @JsonProperty("errors") @JsonPropertyDescription("Error details, which need to be fixed before the workflow can be valid") String errors,
-    @JsonProperty("warnings") @JsonPropertyDescription("Warning details that give additional information") String warnings) {
-}
