+++ server/libs/ai/mcp/mcp-tool/mcp-tool-platform/src/main/resources/instruction_cluster_elements.txt
