+++ server/libs/ai/mcp/mcp-tool/mcp-tool-platform/src/main/resources/instruction_workflow_build.txt
